Architectural update
Totally replaced spherical coordinates with vectors for diffuse materials; massively simplified debugging by separating discrete subpath shading + lens/light gathers from generic subpath connection strategies.

- Fixed a few more hidden bugs; reverse light transport is ~pretty~ now :D

- Checked forward light transport in Shadertoy, found my math/code is fine; system-scale space is just too big to provide enough reflectors for importance probes to reach planets in the scene (and therefore the lens)

- Accelerated camera movement for easier scene navigation

Next update:
- Patch explicit-light-sampling for fractal surfaces
- Patch stochastic wok reduction/denoising
- Separate forward/reverse path reduction
- Separate vertex generation/subpath connection
- Begin implementing MLT/refraction/participating media

+++ Athru/Athru GPU/GPULib/GPULib/PathTracer.cpp
@@ -129,7 +129,7 @@ void PathTracer::PreProcess(const Microsoft::WRL::ComPtr<ID3D11DeviceContext>& c
 	fourByteUnsigned tSecs = std::chrono::duration_cast<std::chrono::duration<fourByteUnsigned>>(currTimeNanoSecs).count();
 	fourByteUnsigned traceableCtr = *(fourByteUnsigned*)(traceableCount.pData);
 	shaderInputPtr->timeDispInfo = DirectX::XMFLOAT4(dt, // Change-in-time in seconds
-													 tSecs, // Current time in seconds
+													 (float)tSecs, // Current time in seconds
 													 (float)traceableCtr, // Number of traceables remaining from the previous frame
 													 GraphicsStuff::DISPLAY_AREA / GraphicsStuff::GROUP_AREA_PATH_REDUCTION); // Total group count (preprocessing)
